change settings form saving
